requirements.txt
altgraph==0.17
asgiref==3.2.10
Django==2.2.14
django-debug-toolbar==2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
fbs==0.8.6
future==0.18.2
macholib==1.14
mysql-connector==2.2.9
packaging==20.4
pefile==2019.4.18
Pillow==6.1.0
PyInstaller==3.4
PyMySQL==0.9.3
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyserial==3.4
pytz==2020.1
pywin32-ctypes==0.2.0
sip==5.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
